Move compute_head_slice_params to common/utils to keep it torch/triton-free
The direct-slice hetero-TP path in nixl/conn.py only needs the integer
head-slice math, but importing it from common/staging_buffer.py dragged the
staging Triton kernels (import torch / import triton at module scope) onto the
non-staging transfer path. Move the helper into common/utils.py, which is pure
integer/numpy code, and import it from there. staging_buffer.py now imports it
from utils, so every existing call site keeps working unchanged.

Behavior is identical: same function body, same results.
